Smoothening out coffee
